+++ spec/models/user_habit_spec.rb
@@ -53,4 +53,3 @@ RSpec.describe UserHabit, type: :model do
     end
   end
 end
-